+++ docs/qubinode/releases.md
@@ -6,6 +6,15 @@
 https://github.com/Qubinode/qubinode-installer/releases
 
 **Qubinode v2.4.4 Release notes**
+* RHEL rhel-8.3-update-2-x86_64-kvm.qcow2 support 
+* RHEL 8.3 Default support
+* OpenShift 4.7.x
+* OKD 4.7.0-0.okd-2021-02-25-144700
+* Added Support for [jig](https://github.com/kenmoini/jig)
+* Force Static IP on IDM server
+
+**Qubinode v2.4.3 Release notes**
+* OpenShift 4.6.x
 * Code improvements to speed up the deployment of OpenShift
 * Deploy the [Kubernetes NFS-Client Provisioner](https://github.com/kubernetes-sigs/nfs-subdir-external-provisioner) by default
 * Configures the Qubinode as a NFS server